~8.15 자잘한 부분 수정.
비디오를 업로드할 경우 hash된 비밀번호를 다시 hash해 로그인이 불가능하던 부분 수정.

비디오를 수정하거나 삭제할때 업로더를 체크하지 않던 부분 수정(프론트에서는 막았지만 백에서는 막지않았음)

+++ src/controller/videoController.js
@@ -30,21 +30,35 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { 
+        user: { _id },
+    } = req.session;
     const video = await Video.findById(id);
     if(!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    // 비디오 업로더와 로그인한 유저가 같지않다면 메인으로 redirect
+    // video.onwer 의 type은 object 이고 _id의 type은 string이기 때문에 데이터형식을 맞춰줘야한다.
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     return res.render("edit", { pageTitle: `Edit ${video.title}`, video });
 }
 
 export const postEdit = async (req, res) => {
     const { id } = req.params;
+    const { 
+        user: { _id },
+    } = req.session;
     //POST 형식으로 보낸 form의 value
     const { title, description, hashtags } = req.body;
     //video object를 받는대신 exists로 필터를 걸어서 true, false를 가져옴
     const video = await Video.exists({_id: id});
     if(!video) {
         return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title,
@@ -88,7 +102,16 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async (req, res) => {
     const { id } = req.params;
-    console.log(id);
+    const { 
+        user: { _id },
+    } = req.session;
+    const video = await Video.findById(id);
+    if(!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
     return res.redirect("/");
 };
